A good amount of work done.
** Plugin system is under way once more. Wondering if we can somehow get a means of generating
specialized plug-in files (like .plgn or .pluginspec) and use that to pull information about a plug-in
without having to going through the hassle of loading every library in the folder and then finding
information about it. The information should be stored in /usr/share/wintermute/plugins for system
plugins (and perhaps $HOME/.config/Wintermute/plugins) as well for user specific plug-ins).

** Now, Wintermute is TTY/GUI agnostic, in a way. At compile-time, you can pick whether or not you
choose to have Wintermute's application core revolve around a GUI (QApplication) or TUI
(QCoreApplication). I wanted to keep it like this so that we'd be able to add Wintermute onto non-