Added script preparing branch for build.
Script named autoprepare.sh is needed because we will generate important
CMake files while building and Core branch is essential for that so we
have to pull everything needed (we have modular branches) and make
proper fixes to files while that.

Another thing done is a first part of CMake files review.
There will be created modules allowing for search of basic Wintermute
plug-ins and libraries (of course, Core being the more prominent example).
In future reviews, new files will be added and draft of CMake support
for foreign project will be done.